@@ -1,21 +1,61 @@
 name: ai_based_farmer_query_app
 
-description: A Flutter application integrating RAG, voice search, image search, and AI capabilities.
+description: A Flutter application integrating RAG, voice search, image search, and AI capabilities for farmer advisory support.
 
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.17.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  rag:
-  voice_search:
-  image_search:
-  ai_integration:
+  cupertino_icons: ^1.0.2
+  http: ^1.2.0
+  provider: ^6.1.2
+  path_provider: ^2.1.3
+  path: ^1.9.0
+  sqflite: ^2.3.0
+  image_picker: ^1.0.8
+  speech_to_text: ^6.3.0
+  camera: ^0.10.5+5
+  flutter_tts: ^3.10.0
+  dio: ^5.6.0
+  flutter_dotenv: ^5.1.0
+  uuid: ^4.3.3
+  intl: ^0.19.0
+  shimmer: ^3.0.0
+  flutter_svg: ^2.0.7
+  get: ^4.6.6
+  flutter_bloc: ^8.1.6
+  equatable: ^2.0.5
+  flutter_spinkit: ^5.2.1
+  connectivity_plus: ^5.0.2
+  permission_handler: ^11.4.0
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+  flutter_lints: ^3.0.0
 
 flutter:
   uses-material-design: true
+  
+  # Assets and fonts
+  assets:
+    - assets/images/
+    - assets/icons/
+    - assets/data/
+    
+  fonts:
+    - family: Poppins
+      fonts:
+        - asset: assets/fonts/Poppins-Regular.ttf
+        - asset: assets/fonts/Poppins-Medium.ttf
+          weight: 500
+        - asset: assets/fonts/Poppins-SemiBold.ttf
+          weight: 600
+        - asset: assets/fonts/Poppins-Bold.ttf
+          weight: 700
